feat : 기존 security 설정 및 사용자 entitiy 삭제(#93)
