--- src/main/java/com/example/servicehub/security/config/CustomComponentConfig.java
@@ -1,103 +1,0 @@
-package com.example.servicehub.security.config;
-
-import com.example.servicehub.repository.ClientRepository;
-import com.example.servicehub.security.SignupManager;
-import com.example.servicehub.security.authentication.CustomOAuth2UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.access.AccessDecisionVoter;
-import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
-import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
-import org.springframework.security.access.vote.RoleHierarchyVoter;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.oauth2.client.OAuth2AuthorizationContext;
-import org.springframework.security.oauth2.client.OAuth2AuthorizeRequest;
-import org.springframework.security.oauth2.client.OAuth2AuthorizedClientProvider;
-import org.springframework.security.oauth2.client.OAuth2AuthorizedClientProviderBuilder;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.client.web.DefaultOAuth2AuthorizedClientManager;
-import org.springframework.security.oauth2.client.web.OAuth2AuthorizedClientRepository;
-import org.springframework.security.oauth2.core.endpoint.OAuth2ParameterNames;
-import org.springframework.util.StringUtils;
-
-import javax.servlet.http.HttpServletRequest;
-import java.time.Duration;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.function.Function;
-
-@Configuration
-@RequiredArgsConstructor
-public class CustomComponentConfig {
-
-    private final ClientRegistrationRepository clientRegistrationRepository;
-    private final OAuth2AuthorizedClientRepository authorizedClientRepository;
-    private final ClientRepository clientRepository;
-
-    @Bean
-    public PasswordEncoder passwordEncoder(){
-        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-    }
-
-    @Bean
-    public SignupManager signupManager(){
-        return new SignupManager();
-    }
-
-    @Bean
-    public CustomOAuth2UserService customOAuth2UserService(){
-        return new CustomOAuth2UserService(clientRepository);
-    }
-
-    @Bean
-    public DefaultOAuth2AuthorizedClientManager authorizedClientManager() {
-
-        OAuth2AuthorizedClientProvider authorizedClientProvider =
-                OAuth2AuthorizedClientProviderBuilder.builder()
-                        .authorizationCode()
-                        .clientCredentials()
-                        .password(passwordGrantBuilder -> passwordGrantBuilder.clockSkew(Duration.ofSeconds(3600)))
-                        .refreshToken(refreshTokenGrantBuilder -> refreshTokenGrantBuilder.clockSkew(Duration.ofSeconds(3600)))
-                        .build();
-
-        DefaultOAuth2AuthorizedClientManager authorizedClientManager =
-                new DefaultOAuth2AuthorizedClientManager(
-                        clientRegistrationRepository, authorizedClientRepository);
-
-        authorizedClientManager.setAuthorizedClientProvider(authorizedClientProvider);
-        authorizedClientManager.setContextAttributesMapper(contextAttributesMapper());
-
-        return authorizedClientManager;
-    }
-
-    private Function<OAuth2AuthorizeRequest, Map<String, Object>> contextAttributesMapper() {
-        return authorizeRequest -> {
-            Map<String, Object> contextAttributes = Collections.emptyMap();
-            HttpServletRequest servletRequest = authorizeRequest.getAttribute(HttpServletRequest.class.getName());
-            String username = servletRequest.getParameter(OAuth2ParameterNames.USERNAME);
-            String password = servletRequest.getParameter(OAuth2ParameterNames.PASSWORD);
-            if (StringUtils.hasText(username) && StringUtils.hasText(password)) {
-                contextAttributes = new HashMap<>();
-                contextAttributes.put(OAuth2AuthorizationContext.USERNAME_ATTRIBUTE_NAME, username);
-                contextAttributes.put(OAuth2AuthorizationContext.PASSWORD_ATTRIBUTE_NAME, password);
-            }
-            return contextAttributes;
-        };
-    }
-
-    @Bean
-    public RoleHierarchy roleHierarchy(){
-        RoleHierarchyImpl roleHierarchy = new RoleHierarchyImpl();
-        roleHierarchy.setHierarchy("ROLE_ADMIN > ROLE_USER");
-        return roleHierarchy;
-    }
-
-    @Bean
-    public AccessDecisionVoter<? extends Object> roleVoter(){
-        return new RoleHierarchyVoter(roleHierarchy());
-    }
-
-}
