--- src/main/resources/templates/client/signin.html
@@ -1,103 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"  xmlns:th="http://www.thymeleaf.org">
-
-<head>
-    <div th:insert="~{fragment/header :: header}"></div>
-    <!-- Custom fonts for this template-->
-    <link href="vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
-    <link
-            href="https://fonts.googleapis.com/css?family=Nunito:200,200i,300,300i,400,400i,600,600i,700,700i,800,800i,900,900i"
-            rel="stylesheet">
-
-    <!-- Custom styles for this template-->
-    <link href="css/sb-admin-2.min.css" rel="stylesheet">
-
-</head>
-<style>
-    .field-error {
-        border-color: #dc3545;
-        color: #dc3545;
-    }
-</style>
-<body class="">
-
-<div class="container" >
-
-    <!-- Outer Row -->
-    <div class="row justify-content-center">
-
-        <div class="col-xl-10 col-lg-12 col-md-9" >
-
-            <div class="card o-hidden border-0 shadow-lg my-5" style="width: 100%">
-                <div class="card-body p-0" >
-                    <!-- Nested Row within Card Body -->
-                    <div class="row justify-content-center">
-                        <div class="col-lg-6 ">
-                            <div class="p-5 ">
-                                <div class="text-center">
-                                    <h1 class="h4 text-gray-900 mb-4">Welcome Back!</h1>
-                                </div>
-                                <form th:action="@{/password/keycloak/login}" method="post">
-                                    <div class="form-group">
-                                        <input type="text" class="form-control form-control-user"
-                                               id="exampleInputEmail" aria-describedby="emailHelp"
-                                               placeholder="Enter username" name="username">
-                                    </div>
-                                    <div class="form-group">
-                                        <input type="password" class="form-control form-control-user"
-                                               id="exampleInputPassword" name="password" placeholder="Password">
-                                    </div>
-                                    <small class="field-error" th:if="${error}" th:text="|아이디 혹은 비밀번호가 맞지 않습니다.|" ></small>
-                                    <div class="form-group">
-                                        <div class="custom-control custom-checkbox small">
-                                            <input type="checkbox" class="custom-control-input" id="customCheck">
-                                            <label class="custom-control-label" for="customCheck">Remember
-                                                Me</label>
-                                        </div>
-                                    </div>
-                                    <button type="submit" class="btn btn-primary btn-user btn-block">
-                                        Login
-                                    </button>
-                                </form>
-                                <hr>
-                                    <div class="text-center">
-                                        <button style="height: 69px" th:onclick="|location.href='@{/oauth2/authorization/naver}'|"class="btn">
-                                            <img src="naver3.png" width="45px">
-                                        </button>
-                                        <button style="height: 69px" th:onclick="|location.href='@{/oauth2/authorization/google}'|"class="btn">
-                                             <img src="google5.png" width="45px">
-                                        </button>
-                                        <button style="height: 69px" th:onclick="|location.href='@{/oauth2/authorization/kakao}'|"class="btn">
-                                             <img src="kakao_login_medium.png">
-                                        </button>
-                                    </div>
-                                <hr>
-                                <div class="text-center">
-                                    <a class="small" th:href="@{/signup}">Create an Account!</a>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-        </div>
-
-    </div>
-
-</div>
-
-<div th:insert="~{fragment/library :: libraries}"></div>
-<!-- Bootstrap core JavaScript-->
-<script src="vendor/jquery/jquery.min.js"></script>
-<script src="vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
-
-<!-- Core plugin JavaScript-->
-<script src="vendor/jquery-easing/jquery.easing.min.js"></script>
-
-<!-- Custom scripts for all pages-->
-<script src="js/sb-admin-2.min.js"></script>
-
-</body>
-
-</html>
