--- src/main/resources/templates/client/signup.html
@@ -1,97 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"  xmlns:th="http://www.thymeleaf.org">
-
-<head>
-
-    <div th:insert="~{fragment/header :: header}"></div>
-    <title>SB Admin 2 - Register</title>
-
-    <!-- Custom fonts for this template-->
-    <link href="vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
-    <link
-            href="https://fonts.googleapis.com/css?family=Nunito:200,200i,300,300i,400,400i,600,600i,700,700i,800,800i,900,900i"
-            rel="stylesheet">
-
-    <!-- Custom styles for this template-->
-    <link href="css/sb-admin-2.min.css" rel="stylesheet">
-
-</head>
-
-<style>
-    .field-error {
-        border-color: #dc3545;
-        color: #dc3545;
-    }
-</style>
-
-<body class="">
-
-<div class="container"  style="width: 60%;display: flex;flex-direction: column;align-items: center;justify-content: center">
-
-    <div class="card shadow-lg my-5" style="width: 50%;display: flex;flex-direction: column;align-items: center;justify-content: center">
-        <div class="card-body" style="width: 100%;">
-            <!-- Nested Row within Card Body -->
-            <div class=" text-center ">
-                <div class="col text-center">
-
-                    <div class="text-center">
-                        <h1 class="h4 text-gray-900 mb-4">Create an Account!</h1>
-                    </div>
-                    <form class="user" th:action="@{/signup}" th:object="${signupForm}" method="post">
-                        <div class="form-group row">
-                            <div class="col-sm-12 mb-3 mb-sm-0">
-                                <input type="text" class="form-control form-control-user" th:field="*{username}" name="username" id="exampleFirstName"
-                                       placeholder="First Name">
-                                <small class="field-error" th:errors="*{username}"></small>
-                            </div>
-                        </div>
-                        <div class="form-group">
-                            <input type="email" class="form-control form-control-user" name="email" th:field="*{email}" id="exampleInputEmail"
-                                   placeholder="Email Address">
-                            <small class="field-error" th:errors="*{email}"></small>
-                        </div>
-                        <div class="form-group row">
-                            <div class="col-sm-6 mb-3 mb-sm-0">
-                                <input type="password" class="form-control form-control-user"
-                                       id="exampleInputPassword" name="password" th:field="*{password}" placeholder="Password">
-
-                                <small class="field-error" th:errors="*{password}"></small>
-                            </div>
-                            <div class="col-sm-6">
-                                <input type="password" class="form-control form-control-user"
-                                       id="exampleRepeatPassword" name="repeatPassword" th:field="*{repeatPassword}" placeholder="Repeat Password">
-                                <small class="field-error" th:errors="*{repeatPassword}"></small>
-                            </div>
-                            <small class="form-text text-muted pt-1"><i class="fas fa-info-circle me-2 text-primary"></i>비밀번호는 8자 이상이어야 하며, 숫자/대문자/소문자/특수문자를 모두 포함해야 합니다.</small>
-                        </div>
-                        <button type="submit" class="btn btn-primary btn-user btn-block">
-                            Register Account
-                        </button>
-
-                    </form>
-                    <hr>
-                    <div class="text-center">
-                        <a class="small" th:href="@{/login}">Already have an account? Login!</a>
-                    </div>
-
-                </div>
-            </div>
-        </div>
-    </div>
-
-</div>
-
-<div th:insert="~{fragment/library :: libraries}"></div>
-<!-- Bootstrap core JavaScript-->
-<script src="/vendor/jquery/jquery.min.js"></script>
-<script src="/vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
-
-<!-- Core plugin JavaScript-->
-<script src="/vendor/jquery-easing/jquery.easing.min.js"></script>
-
-<!-- Custom scripts for all pages-->
-<script src="/js/sb-admin-2.min.js"></script>
-
-</body>
-
-</html>
